core: code cleanup & improvement
