+++ .prettierrc
@@ -1,9 +1,5 @@
 {
-  "plugins": [
-    "prettier-plugin-tailwindcss",
-    "prettier-plugin-sort-imports",
-    "prettier-plugin-prisma"
-  ],
+  "plugins": ["prettier-plugin-tailwindcss", "prettier-plugin-sort-imports"],
   "trailingComma": "none",
   "tabWidth": 2,
   "semi": false,
