--- components/AdBanner.tsx
@@ -1,46 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-
-interface AdBannerProps {
-  adClient: string
-  adSlot: string
-  adFormat?: string
-  fullWidthResponsive?: boolean
-}
-
-declare global {
-  interface Window {
-    adsbygoogle: any[]
-  }
-}
-
-const AdBanner: React.FC<AdBannerProps> = ({
-  adClient,
-  adSlot,
-  adFormat = 'auto',
-  fullWidthResponsive = true
-}) => {
-  useEffect(() => {
-    try {
-      if (typeof window !== 'undefined') {
-        ;(window.adsbygoogle = window.adsbygoogle || []).push({})
-      }
-    } catch (e) {
-      console.error('Adsense error', e)
-    }
-  }, [])
-
-  return (
-    <ins
-      className="adsbygoogle"
-      style={{ display: 'block' }}
-      data-ad-client={adClient}
-      data-ad-slot={adSlot}
-      data-ad-format={adFormat}
-      data-full-width-responsive={fullWidthResponsive ? 'true' : 'false'}
-    />
-  )
-}
-
-export default AdBanner
