+++ data/index.ts
@@ -1,0 +1,3 @@
+export { projects, contributions } from './projects'
+export { knowledge } from './knowledge'
+export { contacts } from './contacts'
