--- helpers/resendClient.ts
@@ -1,5 +1,0 @@
-import { Resend } from 'resend'
-
-const resend = new Resend(process.env.RESEND_API)
-
-export default resend
