+++ interfaces/contacts.ts
@@ -1,0 +1,7 @@
+import { ElementType, SVGProps } from 'react'
+
+export interface Contact {
+  label: string
+  icon: ElementType<SVGProps<SVGSVGElement>>
+  href: string
+}
