+++ interfaces/index.ts
@@ -1,0 +1,3 @@
+export type { Project, Contribution } from './projects'
+export type { Knowledge } from './knowledge'
+export type { Contact } from './contacts'
