+++ package.json
@@ -8,68 +8,43 @@
     "start": "next start",
     "lint": "next lint",
     "format": "prettier --write .",
-    "prisma:generate": "prisma generate",
-    "prisma:migrate": "prisma migrate dev --name",
-    "prisma:push": "prisma db push",
-    "postinstall": "npm run prisma:generate",
     "postbuild": "next-sitemap"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.6.0",
-    "@prisma/client": "^5.17.0",
-    "@radix-ui/react-dialog": "^1.1.0",
-    "@radix-ui/react-dropdown-menu": "^2.1.0",
-    "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-dropdown-menu": "^2.1.1",
     "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-toast": "^1.2.0",
-    "@radix-ui/react-tooltip": "^1.1.0",
-    "@react-email/components": "^0.0.19",
-    "@react-email/render": "^0.0.15",
-    "@supabase/ssr": "^0.4.0",
-    "@supabase/supabase-js": "^2.43.5",
-    "@types/html-to-text": "^9.0.4",
+    "@radix-ui/react-toast": "^1.2.1",
+    "@radix-ui/react-tooltip": "^1.1.2",
     "@vercel/analytics": "^1.3.1",
     "@vercel/speed-insights": "^1.0.12",
-    "bcrypt": "^5.1.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "framer-motion": "^11.2.11",
-    "lucide-react": "^0.395.0",
+    "framer-motion": "^11.5.4",
+    "lucide-react": "^0.439.0",
     "mini-svg-data-uri": "^1.4.4",
-    "next": "^14.2.5",
+    "next": "^14.2.9",
     "next-sitemap": "^4.2.3",
     "next-themes": "^0.3.0",
-    "prisma": "^5.17.0",
     "radix-ui": "^1.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.52.0",
-    "react-responsive-masonry": "^2.2.0",
-    "react-router-dom": "^6.23.1",
-    "react-swipeable": "^7.0.1",
-    "resend": "^3.3.0",
-    "styled-components": "^6.1.11",
-    "supabase": "^1.178.2",
-    "swr": "^2.2.5",
-    "tailwind-merge": "^2.3.0",
-    "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.23.8"
+    "react-router-dom": "^6.26.2",
+    "tailwind-merge": "^2.5.2",
+    "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@types/bcrypt": "^5.0.2",
-    "@types/node": "^20.14.6",
-    "@types/react": "^18.3.3",
+    "@types/node": "^20.16.5",
+    "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
-    "@types/react-responsive-masonry": "^2.1.3",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^9.5.0",
-    "eslint-config-next": "^14.2.4",
-    "postcss": "^8.4.38",
-    "prettier": "^3.3.2",
-    "prettier-plugin-prisma": "^5.0.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "14.2.9",
+    "eslint-plugin-import": "2.30.0",
+    "postcss": "^8.4.45",
+    "prettier": "^3.3.3",
     "prettier-plugin-sort-imports": "^1.8.6",
-    "prettier-plugin-tailwindcss": "^0.6.5",
+    "prettier-plugin-tailwindcss": "^0.6.6",
     "tailwindcss": "3.4.4",
-    "typescript": "^5.4.5"
+    "typescript": "^5.6.2"
   }
 }
