--- prisma/schema.prisma
@@ -1,17 +1,0 @@
-datasource db {
-  provider  = "postgresql"
-  url       = env("DATABASE_URL")
-  directUrl = env("DIRECT_URL")
-}
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-model Reviews {
-  id       Int     @id @default(autoincrement())
-  name     String
-  content  String
-  approved Boolean @default(false)
-  ip       String
-}
